requirements.txt
fastapi==0.109.1
uvicorn==0.27.1
python-dotenv==1.0.0
groq==0.4.2
pydantic==2.6.1
python-multipart==0.0.9